requirements/base.txt
Django==1.10.0
djangorestframework==3.4.3
django-cors-headers==1.1.0
django-filter==0.13.0
django-storages==1.5.0
django-markwhat==1.5.0
djangorestframework-camel-case==0.2.0
docutils==0.12.0
Pillow==3.3.0
psycopg2==2.6.2
pytz==2016.6.1
boto==2.42.0
semantic-version==2.5.0
